+++ src/FSharp.Analyzers.SDK/FSharp.Analyzers.SDK.fs
@@ -148,6 +148,7 @@ type Fix =
         ToText: string
     }
 
+[<RequireQualifiedAccess>]
 type Severity =
     | Info
     | Hint